refactor(agent): extract stress-roll-dialog module
Move buildStressRollDialog out of AgentSheet.ts into its own
module to reduce sheet file size.

refs #559
